#68 Move router to lib/.
