+++ lib/router.js
